requirements.txt
pytorch==1.10
nemo-toolkit[ASR]==1.5.1
